proyecthtml/requirements.txt
Django==5.0.7
django-debug-toolbar==4.4.6
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
django-thumbnails-0.8.0